+++ UPGRADE.md
@@ -1,26 +1,48 @@
 # Upgrading to 4.0.0
 
-## Constructors Of `FetchServer` and `Server` Take An Object
+## Renamed `Server` And `FetchServer`
 
-For `Server`:
+The `Server` class has been renamed to `SinonServer`.
 
 ```diff
-import { Server } from 'fakerest';
-import { data } from './data';
+-import { Server } from 'fakerest';
++import { SinonServer } from 'fakerest';
 
 -const server = new Server('http://myapi.com');
-+const server = new Server({ baseUrl: 'http://myapi.com' });
++const server = new SinonServer({ baseUrl: 'http://myapi.com' });
+```
+
+The `FetchServer` class has been renamed to `FetchMockServer`.
+
+```diff
+-import { FetchServer } from 'fakerest';
++import { FetchMockServer } from 'fakerest';
+
+-const server = new FetchServer('http://myapi.com');
++const server = new FetchMockServer({ baseUrl: 'http://myapi.com' });
+```
+
+## Constructors Of `SinonServer` and `FetchMockServer` Take An Object
+
+For `SinonServer`:
+
+```diff
+import { SinonServer } from 'fakerest';
+import { data } from './data';
+
+-const server = new SinonServer('http://myapi.com');
++const server = new SinonServer({ baseUrl: 'http://myapi.com' });
 server.init(data);
 ```
 
 For `FetchServer`:
 
 ```diff
-import { FetchServer } from 'fakerest';
+import { FetchMockServer } from 'fakerest';
 import { data } from './data';
 
--const server = new FetchServer('http://myapi.com');
-+const server = new FetchServer({ baseUrl: 'http://myapi.com' });
+-const server = new FetchMockServer('http://myapi.com');
++const server = new FetchMockServer({ baseUrl: 'http://myapi.com' });
 server.init(data);
 ```
